worker/pyproject.toml
[project]
name = "worker"
version = "1.0.0"
description = "Worker"
requires-python = ">=3.11,<3.13"
dependencies = [
    "grpcio>=1.76.0",
    "grpcio-tools>=1.76.0",
]